+++ Editor/ValidationSuite/BundledPackagesChecker.cs
@@ -9,7 +9,12 @@ namespace UnityEditor.PackageManager.ValidationSuite
     [UsedImplicitly]
     class BundledPackagesChecker : IPvpChecker
     {
-        public string[] Checks { get; } = { "PVP-163-1" }; // Dependencies of bundled packages should also be bundled
+        internal static readonly string[] StaticChecks = {
+            "PVP-163-1", // Dependencies of bundled packages should also be bundled
+            "PVP-163-2", // Editor manifest consistency checks for dependencies
+        };
+
+        public string[] Checks => StaticChecks;
 
         static readonly string k_EditorManifestPath = EditorApplication.applicationContentsPath + "/Resources/PackageManager/Editor/manifest.json";
 
@@ -17,13 +22,34 @@ namespace UnityEditor.PackageManager.ValidationSuite
 
         public void Run(in PvpRunner.Input input, PvpRunner.Output output)
         {
+            Json editorManifest;
+            int schemaVersion;
             try
             {
-                var editorManifest = new Json(XrayUtils.DecodeUtf8Lax(ReadAllBytes(k_EditorManifestPath)), null);
+                editorManifest = new Json(XrayUtils.DecodeUtf8Lax(ReadAllBytes(k_EditorManifestPath)), null);
+                schemaVersion = (int)(editorManifest["schemaVersion"].IfPresent?.Number ?? 1.0);
+            }
+            catch (IOException e) when (e.IsNotFoundError())
+            {
+                foreach (var check in StaticChecks)
+                    output.Error(check, "editor manifest not found");
+                return;
+            }
+            catch (SimpleJsonException)
+            {
+                foreach (var check in StaticChecks)
+                    output.Error(check, "invalid editor manifest");
+                return;
+            }
 
-                var schemaVersion = (int)(editorManifest["schemaVersion"].IfPresent?.Number ?? 1.0);
-                if (schemaVersion < 4) return;
+            if (schemaVersion >= 4) CheckV1(input, output, editorManifest);
+            CheckV2(input, output, editorManifest);
+        }
 
+        static void CheckV1(in PvpRunner.Input input, PvpRunner.Output output, Json editorManifest)
+        {
+            try
+            {
                 var packageUnderTest = input.Package.name;
                 if (!MustBeBundled(packageUnderTest)) return;
 
@@ -46,13 +72,43 @@ namespace UnityEditor.PackageManager.ValidationSuite
                     return entry["mustBeBundled"].IfPresent?.Boolean ?? entry["version"].IsPresent;
                 }
             }
-            catch (IOException e) when (e.IsNotFoundError())
-            {
-                output.Error("PVP-163-1", "editor manifest not found");
-            }
             catch (SimpleJsonException)
             {
                 output.Error("PVP-163-1", "invalid editor manifest");
+            }
+        }
+
+        static void CheckV2(in PvpRunner.Input input, PvpRunner.Output output, Json editorManifest)
+        {
+            DependencyConsistencyVerifier.EditorManifest virtualEditorManifest;
+            try
+            {
+                virtualEditorManifest = DependencyConsistencyVerifier.EditorManifest.Parse(editorManifest);
+            }
+            catch (SimpleJsonException)
+            {
+                output.Error("PVP-163-2", "invalid editor manifest");
+                return;
+            }
+
+            // Update the "virtual" editor manifest with information from the publish set.
+            foreach (var p in input.PublishSet)
+            {
+                if (virtualEditorManifest.Entries.TryGetValue(p.Id.Name, out var entry))
+                {
+                    entry.Version = p.Id.Version;
+                    virtualEditorManifest.Entries[p.Id.Name] = entry;
+                }
+            }
+
+            try
+            {
+                var packageManifest = new Json(XrayUtils.DecodeUtf8Lax(input.Manifest), "package.json");
+                DependencyConsistencyVerifier.EditorManifestCheck(packageManifest, virtualEditorManifest, e => output.Error("PVP-163-2", e));
+            }
+            catch (SimpleJsonException e)
+            {
+                output.Error("PVP-163-2", e.FullMessage);
             }
         }
     }