+++ Lib/PvpXray/StringSlice.cs
@@ -19,7 +19,10 @@ namespace PvpXray
         public bool Equals(StringSlice other) => Length == other.Length && string.CompareOrdinal(String, Start, other.String, other.Start, Length) == 0;
         public bool EqualsIgnoreCase(StringSlice other) => Length == other.Length && string.Compare(String, Start, other.String, other.Start, Length, StringComparison.OrdinalIgnoreCase) == 0;
 
+        public StringSlice Slice(int start) => new StringSlice { String = String, Start = Start + start, End = End };
         public StringSlice Slice(int start, int end) => new StringSlice { String = String, Start = Start + start, End = Start + end };
+
+        public bool StartsWith(StringSlice value) => Length >= value.Length && string.CompareOrdinal(String, Start, value.String, value.Start, value.Length) == 0;
 
         public bool TryIndexOf(char c, out int i)
         {
@@ -27,6 +30,14 @@ namespace PvpXray
             if (i == -1) return false;
             i -= Start;
             return true;
+        }
+
+        public StringSlice Trim(params char[] trimChars)
+        {
+            int start = Start, end = End;
+            while (start < end && Array.IndexOf(trimChars, String[start]) != -1) start++;
+            while (start < end && Array.IndexOf(trimChars, String[end - 1]) != -1) end--;
+            return new StringSlice { String = String, Start = start, End = end };
         }
 
         public override string ToString() => String.Substring(Start, Length);